cast positions dtype in XGLMModel

+++ src/transformers/models/xglm/modeling_xglm.py
@@ -709,8 +709,8 @@ class XGLMModel(XGLMPreTrainedModel):
             # [bsz, seq_len] -> [bsz, 1, tgt_seq_len, src_seq_len]
             encoder_attention_mask = _expand_mask(encoder_attention_mask, inputs_embeds.dtype, tgt_len=input_shape[-1])
 
-        # embed positions
-        positions = self.embed_positions(input_ids, inputs_embeds, past_key_values_length)
+        # embed positions, cast from float32 to `inputs_embeds.dtype`
+        positions = self.embed_positions(input_ids, inputs_embeds, past_key_values_length).to(inputs_embeds.dtype)
 
         hidden_states = inputs_embeds + positions
